# PERF: avoid copy in replace